Implement the user reputation system.
How?
When a question/answer/comment is upvoted, 5 points are added to the `points` attribute of the user that made the said resource.
If it is a downvote instead, 2 is subtracted from the that user
NB: If the same user upvotes and downvotes a resource, the previous option is canceled and the new one takes affect as though nothing happened.